Windows: Refactor volumes.go for future Windows Daemon.
