+++ daemon/volumes.go
@@ -8,7 +8,6 @@ import (
 	"path/filepath"
 	"sort"
 	"strings"
-	"syscall"
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/docker/docker/daemon/execdriver"
@@ -36,15 +35,6 @@ func (mnt *Mount) Export(resource string) (io.ReadCloser, error) {
 		return nil, err
 	}
 	return mnt.volume.Export(path, name)
-}
-
-func (container *Container) prepareVolumes() error {
-	if container.Volumes == nil || len(container.Volumes) == 0 {
-		container.Volumes = make(map[string]string)
-		container.VolumesRW = make(map[string]bool)
-	}
-
-	return container.createVolumes()
 }
 
 // sortedVolumeMounts returns the list of container volume mount points sorted in lexicographic order
@@ -380,20 +370,4 @@ func copyExistingContents(source, destination string) error {
 	}
 
 	return copyOwnership(source, destination)
-}
-
-// copyOwnership copies the permissions and uid:gid of the source file
-// into the destination file
-func copyOwnership(source, destination string) error {
-	var stat syscall.Stat_t
-
-	if err := syscall.Stat(source, &stat); err != nil {
-		return err
-	}
-
-	if err := os.Chown(destination, int(stat.Uid), int(stat.Gid)); err != nil {
-		return err
-	}
-
-	return os.Chmod(destination, os.FileMode(stat.Mode))
 }
